Refactor: BoxOfficeView 분리 (#12)
